refactor: Move login, register and logout functionality to Home module
- Moved login, register and logout methods from UserController and UserService to HomeController and HomeService.
- Created a new HomeModule to encapsulate the HomeController and HomeService, including necessary imports and TypeORM integration.
- Updated related tests by moving login, register and logout tests to home.e2e-spec and adjusting UserController e2e tests accordingly.
- Adjusted endpoints from `/home` to `/` for HomeController routes, including testing in Postman and the browser to verify functionality.
