+++ src/app.module.ts
@@ -1,13 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { DatabaseModule } from './database/database.module';
-import { HomeController } from './home.controller';
 import { Goal } from './entities/goal.entity';
 import { Interval } from './entities/interval.entity';
 import { User } from './entities/user.entity';
-import { HomeService } from './home.service';
 import { UserModule } from './user/user.module';
 import { IntervalModule } from './interval/interval.module';
 import { GoalModule } from './goal/goal.module';
@@ -15,6 +11,7 @@ import { roles } from './roles/roles';
 import { AccessControlModule } from 'nest-access-control';
 import { ThrottlerGuard, ThrottlerModule } from '@nestjs/throttler';
 import { APP_GUARD } from '@nestjs/core';
+import { HomeModule } from './home/home.module';
 
 @Module({
   imports: [
@@ -24,6 +21,7 @@ import { APP_GUARD } from '@nestjs/core';
     UserModule,
     IntervalModule,
     GoalModule,
+    HomeModule,
     ThrottlerModule.forRoot([
       {
         ttl: 60000,
@@ -31,10 +29,7 @@ import { APP_GUARD } from '@nestjs/core';
       },
     ]),
   ],
-  controllers: [AppController, HomeController],
   providers: [
-    AppService,
-    HomeService,
     {
       provide: APP_GUARD,
       useClass: ThrottlerGuard,
