+++ src/home/home.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { HomeController } from './home.controller';
+import { HomeService } from './home.service';
+import { User } from '../entities/user.entity';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([User])],
+  controllers: [HomeController],
+  providers: [HomeService],
+  exports: [HomeService],
+})
+export class HomeModule {}
