+++ src/views/partials/footer.ejs
@@ -2,13 +2,13 @@
   <div>
     <p>&copy; <%= new Date().getFullYear() %> Mindful Zen. All rights reserved.</p>
     <nav>
-      <a href="/home">Home</a>
+      <a href="/">Home</a>
       <% if (currentUser) { %>
         <a href="/profile">Profile</a>
-        <a href="/users/logout">Logout</a>
+        <a href="/logout">Logout</a>
       <% } else { %>
-        <a href="/users/login">Login</a>
-        <a href="/users/register">Register</a>
+        <a href="/login">Login</a>
+        <a href="/register">Register</a>
       <% } %>
       <a href="/about">About Us</a>
       <a href="/contact">Contact</a>
