+++ src/views/partials/header.ejs
@@ -3,10 +3,10 @@
     <p>Welcome to Mindful Zen!</p>
     <% if (currentUser) { %>
       <p>Welcome, <%= currentUser.name ? currentUser.name : 'User' %>!</p>
-      <a href="/users/logout">Logout</a>
+      <a href="/logout">Logout</a>
     <% } else { %>
-      <a href="/users/login">Login</a>
-      <a href="/users/register">Register</a>
+      <a href="/login">Login</a>
+      <a href="/register">Register</a>
     <% } %>
   </nav>
 </header>
